+++ extensions/arc/src/actions.tsx
@@ -313,3 +313,14 @@ export function EditTabActionSection(props: { tab: Tab; mutate: MutatePromise<Ta
     </ActionPanel.Section>
   );
 }
+
+export function CreateQuickLinkActionSection(props: { url: string; title?: string }) {
+  return (
+    <ActionPanel.Section>
+      <Action.CreateQuicklink
+        quicklink={{ link: props.url, name: props.title, application: "Arc" }}
+        shortcut={{ modifiers: ["cmd"], key: "s" }}
+      />
+    </ActionPanel.Section>
+  );
+}