+++ extensions/gitlab/src/components/commits/list.tsx
@@ -54,7 +54,9 @@ function EventCommitListItem(props: { event: Event }): JSX.Element {
     return null;
   };
 
-  const statusIcon: Image.ImageLike | undefined = status?.status ? getCIJobStatusIcon(status.status) : undefined;
+  const statusIcon: Image.ImageLike | undefined = status?.status
+    ? getCIJobStatusIcon(status.status, status.allow_failure)
+    : undefined;
   const icon: Image.ImageLike | undefined = statusIcon
     ? statusIcon
     : { source: GitLabIcons.commit, tintColor: Color.Green };
@@ -115,6 +117,7 @@ export interface CommitStatus {
   status: string;
   author: User;
   ref?: string;
+  allow_failure: boolean;
 }
 
 export interface Commit {