+++ extensions/internet-radio/src/stop-playback.ts
@@ -6,9 +6,16 @@ export default async function Command() {
   const currentStationName = await LocalStorage.getItem("-current-station-name");
   const currentTrackID = await LocalStorage.getItem("-current-track-id");
 
-  if (currentStationName && currentTrackID && currentStationName != "" && currentTrackID != "") {
+  if (currentTrackID && currentTrackID != "") {
     await pausePlayback();
     await deleteTrack(currentTrackID as string);
+
+    if (environment.launchType == LaunchType.UserInitiated) {
+      await showToast({ title: "Stopped Station", message: currentStationName as string });
+    }
+  } else if (currentStationName && currentStationName != "") {
+    await pausePlayback();
+    await deleteTrack(undefined, `Raycast: ${currentStationName}`);
 
     if (environment.launchType == LaunchType.UserInitiated) {
       await showToast({ title: "Stopped Station", message: currentStationName as string });