+++ extensions/laravel-valet/src/types/entities.d.ts
@@ -1,0 +1,13 @@
+export declare type Config = {
+  paths: string[];
+  tld: string;
+  loopBack: string;
+};
+
+export declare type Site = {
+  name: string;
+  path: string;
+  prettyPath: string;
+  secured: boolean;
+  url: string;
+};